requirements/readthedocs.txt
albumentations
imgaug
kwarray
lanms-proper
lmdb
matplotlib
mmcv
Polygon3
pyclipper
rapidfuzz
fvcore
regex
scikit-image
scipy
shapely
titlecase
torch
torchvision
